+++ charlotte-third-places/components/PlaceModal.tsx
@@ -3,19 +3,13 @@
 import { Place } from "@/lib/types";
 import { Icons } from "@/components/Icons";
 import { Button } from "@/components/ui/button";
-import { Separator } from "@/components/ui/separator";
-import { ShareButton } from "@/components/ShareButton";
-import { QuickFacts } from "@/components/QuickFacts";
+import { PlaceContent } from "@/components/PlaceContent";
 import {
     FC,
     useRef,
-    useEffect,
-    useMemo
+    useEffect
 } from "react";
 import React from "react";
-import { ResponsiveLink } from "@/components/ResponsiveLink";
-import { RichTextSection } from "@/components/RichTextSection";
-import { useModalContext } from "@/contexts/ModalContext";
 import {
     Dialog,
     DialogContent,
@@ -32,7 +26,6 @@ interface PlaceModalProps {
 
 export const PlaceModal: FC<PlaceModalProps> = ({ place, open, onClose }) => {
     const contentRef = useRef<HTMLDivElement>(null);
-    const { showPlacePhotos } = useModalContext();
 
     useEffect(() => {
         // Scroll to the top when the modal opens
@@ -41,29 +34,9 @@ export const PlaceModal: FC<PlaceModalProps> = ({ place, open, onClose }) => {
         }
     }, [open, place]);
 
-    const shareUrl = useMemo(() => {
-        // SSR safety: if `window` is undefined, fallback
-        if (typeof window === "undefined" || !place) {
-            return "";
-        }
-        return `${window.location.origin}/places/${place.recordId}`;
-    }, [place]);
-
     if (!open || !place) {
         return null;
     }
-
-    const hasComments = place.comments?.trim();
-    const hasPhotos = place.photos && place.photos.length > 0;
-    const appleMapsProfileURL = place.appleMapsProfileURL?.trim();
-    const googleMapsProfileURL = place.googleMapsProfileURL?.trim();
-    const website = place.website?.trim();
-    const instagram = place.instagram?.trim();
-    const tiktok = place.tiktok?.trim();
-    const twitter = place.twitter?.trim();
-    const youtube = place.youtube?.trim();
-    const facebook = place.facebook?.trim();
-    const linkedIn = place.linkedIn?.trim();
 
     return (
         <Dialog open={open} onOpenChange={onClose}>
@@ -79,7 +52,6 @@ export const PlaceModal: FC<PlaceModalProps> = ({ place, open, onClose }) => {
                     e.preventDefault();
                 }}
             >
-
                 {/* Featured ribbon, corner banner */}
                 {place.featured && (
                     <div className="absolute top-0 left-0 z-10 overflow-hidden w-44 h-44 pointer-events-none">
@@ -96,92 +68,11 @@ export const PlaceModal: FC<PlaceModalProps> = ({ place, open, onClose }) => {
                     </DialogTitle>
                     <DialogDescription className="text-center">{place.type.join(", ")}</DialogDescription>
                 </DialogHeader>
-                <Separator />
-                {/* Primary Actions */}
-                <div className="space-y-4">
-                    <div className="flex justify-center space-x-2 sm:space-x-4">
-                        {googleMapsProfileURL && (
-                            <ResponsiveLink href={googleMapsProfileURL} aria-label="Visit Google Maps Page">
-                                <Button variant="outline">
-                                    <Icons.google className="h-7 w-7" />
-                                </Button>
-                            </ResponsiveLink>
-                        )}
-                        {appleMapsProfileURL && (
-                            <ResponsiveLink href={appleMapsProfileURL} aria-label="Visit Apple Maps Page">
-                                <Button variant="outline">
-                                    <Icons.apple className="h-7 w-7" />
-                                </Button>
-                            </ResponsiveLink>
-                        )}
-                        {hasPhotos && (
-                            <Button
-                                variant="outline"
-                                onClick={() => showPlacePhotos(place, 'modal')}
-                                aria-label="View photos"
-                            >
-                                <Icons.camera className="h-7 w-7 text-primary" />
-                            </Button>
-                        )}
-                        {website && (
-                            <ResponsiveLink href={website} aria-label="Visit Website">
-                                <Button variant="outline">
-                                    <Icons.globe className="h-7 w-7" />
-                                </Button>
-                            </ResponsiveLink>
-                        )}
-                        <ShareButton
-                            placeName={place.name}
-                            url={shareUrl}
-                            variant="outline"
-                            displayType="icon"
-                            aria-label="Share Place"
-                        />
-                    </div>
-                    <Separator />
-                    <QuickFacts
-                        address={place.address}
-                        neighborhood={place.neighborhood}
-                        size={place.size}
-                        purchaseRequired={place.purchaseRequired}
-                        parking={place.parking}
-                        freeWiFi={place.freeWiFi}
-                        hasCinnamonRolls={place.hasCinnamonRolls}
-                        instagram={instagram}
-                        tiktok={tiktok}
-                        twitter={twitter}
-                        youtube={youtube}
-                        facebook={facebook}
-                        linkedIn={linkedIn}
-                    />
-                    <Separator />
-                    {/* DESCRIPTION - Always visible, high priority */}
-                    <RichTextSection
-                        heading="Description"
-                        priority="high"
-                    >
-                        {place.description?.trim() || "A third place in the Charlotte, North Carolina area."}
-                    </RichTextSection>
 
-                    {/* COMMENTS - Smart truncation for long content */}
-                    {hasComments && (
-                        <>
-                            <Separator />
-                            <RichTextSection
-                                heading="Comments"
-                                priority="medium"
-                            >
-                                {place.comments!}
-                            </RichTextSection>
-                        </>)
-                    }
-                    <Separator className="hidden sm:block" />
-                    <p className="hidden sm:block">
-                        <Icons.folder className="h-4 w-4 text-yellow-400 inline mr-2" />
-                        <span className="font-semibold">Metadata:</span> Added: {new Date(place.createdDate).toLocaleDateString("en-US")} | Last Updated:{" "}
-                        {new Date(place.lastModifiedDate).toLocaleDateString("en-US")}.
-                    </p>
-                </div>
+                <PlaceContent
+                    place={place}
+                    layout="modal"
+                />
 
                 {/* CLOSE BUTTON */}
                 <div className="flex justify-center py-4 px-4 mt-auto">
